moved mod stuff from user-events to a new file because 3000+ lines is stupid
